+++ tests-integration/src/test/java/com/contentful/vaultintegration/SqliteEscapeTest.java
@@ -14,8 +14,8 @@ public class SqliteEscapeTest extends BaseTest {
   }
 
   @Test public void testSync() throws Exception {
-    enqueue("sqlite_escape_space.json");
-    enqueue("sqlite_escape_initial.json");
+    enqueue("escape/space.json");
+    enqueue("escape/initial.json");
     sync();
 
     SqliteEscapeModel item = vault.fetch(SqliteEscapeModel.class).first();