+++ tests-integration/src/test/java/com/contentful/vaultintegration/lib/allthethings/AllTheThingsSpace.java
@@ -1,0 +1,7 @@
+package com.contentful.vaultintegration.lib.allthethings;
+
+import com.contentful.vault.Space;
+
+@Space(value = "hwp8ctxa24x7", models = { AllTheThingsResource.class })
+public class AllTheThingsSpace {
+}