+++ CHANGELOG.md
@@ -3,12 +3,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Version [0.9.5][unreleased] - (in development)
-- TBD
+- Fixed: Boolean field values set properly.
+- Changed: Missing location/object fields will be set to null instead of empty collections.
 
 ## Version [0.9.4] - 2015-06-05
 - Fixed: Replace `getClass()` calls with generated code for link resolution.
 - Fixed: Escape model field names on inserts.
 - Changed: Add `onError` and `onSuccess()` methods to `SyncCallback`.
+- Changed: Use contentful.java v2.0.4.
 
 ## Version [0.9.3] - 2015-05-26
 - New: `Resource` implementation for `equals()` and `hashCode()`